1.多路段运行（不包括信控/非信控交叉口）
2.调试点未取消（四个"BUG"位置，一个时间打印）

3.#开发备忘#

a) 当前设计下 可能会出现部分车进入非可行车道后（di!=0）发现不能通过换道（veh_lane_use）进入可行车道 引发错误 下版本需要改正

b) 当前无法对车道允许运行的车辆类型进行限制，若要加入1.需要完善laneUseRules与车辆类型匹配的mask；2.需要在车道变换前做判断(进入下个seg前(dealPassing)，换道前(dealLC加上判断)，进入下个link前(dealPassing) 以及在CalDi/getValidLanes/pathValid(未实现)中考虑此因素作为行为指引)

c) 下版本需要统一successivelane与connectedDnLane的表达，以及enterallow与分车辆类型允许使用（laneUse）的统一

d) 下版本需要减少CalDi的计算

+++ TranSysLab/src/main/java/com/transyslab/roadnetwork/RoadNetwork.java
@@ -98,7 +98,7 @@ public class RoadNetwork extends SimpleDirectedWeightedGraph<Node, Link>{
 		RoadNetworkPool pool = RoadNetworkPool.getInstance();		
 		if (pool != null) {
 			HashMap<String, Integer> hm = pool.getHashMap();
-			int threadid = hm.get(Thread.currentThread().getName()).intValue();
+			int threadid = hm.get(Thread.currentThread().getName());
 			return RoadNetworkPool.getInstance().getNetwork(threadid);
 		}
 		else {
